WIP - use Symfony's process library to execute tex2svg

+++ src/services/MakePdf.php
@@ -30,6 +30,7 @@ use Psr\Log\NullLogger;
 use function shell_exec;
 use function str_replace;
 use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\Process\Process;
 use function tempnam;
 use function unlink;
 
@@ -187,11 +188,19 @@ class MakePdf extends AbstractMake
         // use tex2svg-page script located in src/node-apps
         // convert tex to svg with nodejs script
         // returns nothing if there is no tex
-        $currentDir = dirname(__DIR__);
-        chdir(dirname(__DIR__, 2) . '/src/node-apps');
-        // disable font cache so all paths are inside the svg and not linked
-        $html = shell_exec('./tex2svg-page --fontCache=none ' . $filename);
-        chdir($currentDir);
+        $process = new Process([
+            dirname(__DIR__, 2) . '/src/node-apps/tex2svg-page',
+            // disable font cache so all paths are inside the svg and not linked
+            '--fontCache=none',
+            $filename
+        ]);
+        $process->run();
+
+        if (!$process->isSuccessful()) {
+            throw new ProcessFailedException($process);
+        }
+
+        $html = $process->getOutput();
         unlink($filename);
 
         // was there actually tex in the content?
